replied to ben's comment

+++ server/test_server.py
@@ -4,16 +4,24 @@ import unittest
 import subprocess
 import socket
 import time
+import platform
+import os
+import signal
 
 
-class Server:
+class Server(object):
     def start(self):
-        self.server = subprocess.Popen(['gunicorn', 'server'], stdin=subprocess.PIPE, stdout=subprocess.PIPE,
-                                       stderr=subprocess.PIPE)
+        if platform.system() == 'Windows':
+            self.server = subprocess.Popen('python server.py', shell=True, stdout=subprocess.PIPE)
+        else:
+            self.server = subprocess.Popen(['gunicorn', 'server'], stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
 
     def shutdown(self):
-        self.server.kill()
-        subprocess.call(['pkill', 'gunicorn'], stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+        if platform.system() == 'Windows':
+            subprocess.call(['taskkill', '/F', '/T', '/PID', str(self.server.pid)])
+        else:
+            self.server.kill()
+            subprocess.call(['pkill', 'gunicorn'], stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
 
     def check_server(self):
         # attempt to connect to localhost/8000
@@ -36,7 +44,7 @@ class Server:
         return status
 
 
-class ServerCommunication:
+class ServerCommunication(object):
     @staticmethod
     def login(payload):
         url = 'http://localhost:8000/api/login'
@@ -47,6 +55,27 @@ class ServerCommunication:
     def register(payload):
         url = 'http://localhost:8000/api/register'
         r = requests.post(url, json=payload)
+        return r.text
+
+    @staticmethod
+    def friendship(header, payload=None):
+        url = 'http://localhost:8000/api/friendship'
+        if payload is not None:
+            r = requests.post(url, json=payload, headers=header)
+        else:
+            r = requests.get(url, headers=header)
+        return r.text
+
+    @staticmethod
+    def friendship_respones(header, payload):
+        url = 'http://localhost:8000/api/friendship/response'
+        r = requests.post(url, json=payload, headers=header)
+        return r.text
+
+    @staticmethod
+    def switch_ip(header, payload):
+        url = 'http://localhost:8000/api/switch-ip'
+        r = requests.get(url, json=payload, headers=header)
         return r.text
 
 
@@ -60,28 +89,48 @@ class TestPostRequest(unittest.TestCase):
             raise Exception('Server could not be started')
 
         # communicating
-        cls.register_send_data = {'username': 'alex', 'password': '1234'}
-        cls.login_send_data = {'username': 'alex', 'password': '1234'}
-        cls.register_ret_data = json.loads(ServerCommunication.login(cls.register_send_data))
+        cls.register_send_data = {'username': 'ben', 'password': '1234'}
+        cls.login_send_data = {'username': 'ben', 'password': '1234'}
+        cls.register_ret_data = json.loads(ServerCommunication.register(cls.register_send_data))
         cls.login_ret_data = json.loads(ServerCommunication.login(cls.login_send_data))
+
+        cls.friendship_send_data = {'reply': 'alex'}
+        cls.friendship_response_send_data = {'reply': 'alex', 'status': 'confirm'}
+        cls.switch_ip_send_data = {'receiver': 'alex'}
+        cls.friendship_ret_data_post = json.loads(ServerCommunication.friendship({'Authorization': cls.login_ret_data['token'].encode('ascii')}, cls.friendship_send_data))
+        cls.friendship_ret_data_get = json.loads(ServerCommunication.friendship({'Authorization': cls.login_ret_data['token'].encode('ascii')}))
+        cls.friendship_response_ret_data = json.loads(ServerCommunication.friendship_respones({'Authorization': cls.login_ret_data['token'].encode('ascii')}, cls.friendship_response_send_data))
+        cls.switch_ip_ret_data = json.loads(ServerCommunication.switch_ip({'Authorization': cls.login_ret_data['token'].encode('ascii')}, cls.switch_ip_send_data))
 
     @classmethod
     def tearDownClass(cls):
         # shutting down the server
+        # alex this (^) is useless comment
+        # like to write on this code:
+        # i++ // increase i by 1
+        # why did you do this ??!!
+
         cls.server.shutdown()
 
-    def testRegisterUsername(self):
+    def test_register_username(self):
         self.assertEquals(self.register_send_data['username'], self.register_ret_data['username'])
 
-    def testRegisterPassword(self):
+    def test_register_password(self):
         self.assertEquals(self.register_send_data['password'], self.register_ret_data['password'])
 
-    def testLoginUsername(self):
+    def test_login_username(self):
         self.assertEquals(self.login_send_data['username'], self.login_ret_data['username'])
 
-    def testLoginPassword(self):
+    def test_login_password(self):
         self.assertEquals(self.login_send_data['password'], self.login_ret_data['password'])
 
+    def test_friendship(self):
+        self.assertEqual(self.friendship_ret_data_post['sender'], self.login_ret_data['token'])
+        self.assertEqual(self.friendship_send_data['reply'], self.friendship_send_data['reply'])
+
+    def test_friendship_request(self):
+        self.assertEqual(self.friendship_response_ret_data['sender'], self.login_ret_data['token'])
+        self.assertEqual(self.friendship_response_send_data['reply'], self.friendship_response_send_data['reply'])
 
 if __name__ == '__main__':
     suite = unittest.TestLoader().loadTestsFromTestCase(TestPostRequest)
